requirements.txt
dash==2.17.1
dash-bootstrap-components==1.6.0
requests==2.32.3
numpy==2.0.1
scikit-learn==1.5.1
python-dotenv==1.0.1